requirements.txt
transformers
datasets
peft
accelerate
streamlit
pyyaml